# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoSphere ESG Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /esg-policies:
    get:
      operationId: listEsgPolicies
      tags: [governance]
      responses:
        "200":
          description: List of ESG policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EsgPolicy"
    post:
      operationId: createEsgPolicy
      tags: [governance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EsgPolicyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsgPolicy"
  /esg-policies/{id}:
    get:
      operationId: getEsgPolicy
      tags: [governance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: ESG Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsgPolicy"
    patch:
      operationId: updateEsgPolicy
      tags: [governance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EsgPolicyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsgPolicy"
    delete:
      operationId: deleteEsgPolicy
      tags: [governance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /policy-acknowledgements:
    get:
      operationId: listPolicyAcknowledgements
      tags: [governance]
      parameters:
        - { name: policyId, in: query, schema: { type: integer } }
        - { name: employeeId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: List of acknowledgements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyAcknowledgement"
    post:
      operationId: createPolicyAcknowledgement
      tags: [governance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyAcknowledgementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyAcknowledgement"
  /audits:
    get:
      operationId: listAudits
      tags: [governance]
      responses:
        "200":
          description: List of audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Audit"
    post:
      operationId: createAudit
      tags: [governance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audit"
  /audits/{id}:
    get:
      operationId: getAudit
      tags: [governance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audit"
    patch:
      operationId: updateAudit
      tags: [governance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audit"
  /compliance-issues:
    get:
      operationId: listComplianceIssues
      tags: [governance]
      parameters:
        - { name: auditId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string } }
        - { name: overdue, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: List of compliance issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceIssue"
    post:
      operationId: createComplianceIssue
      tags: [governance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceIssueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceIssue"
  /compliance-issues/{id}:
    get:
      operationId: getComplianceIssue
      tags: [governance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Compliance issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceIssue"
    patch:
      operationId: updateComplianceIssue
      tags: [governance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceIssueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceIssue"

  # ── Gamification ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    EsgPolicy:
      type: object
      required: [id, title, category, effectiveDate, status]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string, enum: [environmental, social, governance] }
        effectiveDate: { type: string, format: date }
        expiryDate: { type: ["string", "null"] }
        status: { type: string, enum: [draft, active, archived] }
        acknowledgementCount: { type: integer }
        totalEmployees: { type: integer }
    EsgPolicyInput:
      type: object
      required: [title, category, effectiveDate]
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string, enum: [environmental, social, governance] }
        effectiveDate: { type: string, format: date }
        expiryDate: { type: string, format: date }
        status: { type: string, enum: [draft, active, archived], default: draft }
    EsgPolicyUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        effectiveDate: { type: string, format: date }
        expiryDate: { type: string, format: date }
        status: { type: string }

    # ── Policy Acknowledgement ──────────────────────────────────────────────
    PolicyAcknowledgement:
      type: object
      required: [id, policyId, employeeId, acknowledgedAt]
      properties:
        id: { type: integer }
        policyId: { type: integer }
        policyTitle: { type: ["string", "null"] }
        employeeId: { type: integer }
        employeeName: { type: ["string", "null"] }
        acknowledgedAt: { type: string, format: date-time }
    PolicyAcknowledgementInput:
      type: object
      required: [policyId, employeeId]
      properties:
        policyId: { type: integer }
        employeeId: { type: integer }

    # ── Audit ────────────────────────────────────────────────────────────────
    Audit:
      type: object
      required: [id, title, auditor, auditDate, status]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        auditor: { type: string }
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        auditDate: { type: string, format: date }
        status: { type: string, enum: [scheduled, in_progress, completed, cancelled] }
        issueCount: { type: integer }
    AuditInput:
      type: object
      required: [title, auditor, auditDate]
      properties:
        title: { type: string }
        description: { type: string }
        auditor: { type: string }
        departmentId: { type: integer }
        auditDate: { type: string, format: date }
        status: { type: string, enum: [scheduled, in_progress, completed, cancelled], default: scheduled }
    AuditUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        auditor: { type: string }
        auditDate: { type: string, format: date }
        status: { type: string }

    # ── Compliance Issue ────────────────────────────────────────────────────
    ComplianceIssue:
      type: object
      required: [id, auditId, severity, description, ownerId, dueDate, status]
      properties:
        id: { type: integer }
        auditId: { type: integer }
        auditTitle: { type: ["string", "null"] }
        severity: { type: string, enum: [low, medium, high, critical] }
        description: { type: string }
        ownerId: { type: integer }
        ownerName: { type: ["string", "null"] }
        dueDate: { type: string, format: date }
        status: { type: string, enum: [open, in_progress, resolved, overdue] }
        isOverdue: { type: boolean }
        createdAt: { type: string, format: date-time }
    ComplianceIssueInput:
      type: object
      required: [auditId, severity, description, ownerId, dueDate]
      properties:
        auditId: { type: integer }
        severity: { type: string, enum: [low, medium, high, critical] }
        description: { type: string }
        ownerId: { type: integer }
        dueDate: { type: string, format: date }
        status: { type: string, enum: [open, in_progress, resolved, overdue], default: open }
    ComplianceIssueUpdate:
      type: object
      properties:
        severity: { type: string }
        description: { type: string }
        ownerId: { type: integer }
        dueDate: { type: string, format: date }
        status: { type: string }

    # ── Challenge ────────────────────────────────────────────────────────────
...